move righthand panel styles to its own file
